requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.3
sqlalchemy~=2.0.30
aiogram~=3.7.0
pandas~=2.2.2
python-dotenv~=1.0.0
selenium~=4.21.0